Close to MVP for home page done, need to work on about page
I have completed the JS class and constructor for my home page. As far as media querries go, the nav bar, the images for teh student reviews, adn the height of each section of the page needs to be adjsuted. I have completed my footer on both the landing page and the about page that uses two ecent listeners each - though they are the same ones. My student reviews use a JS constructor. If I am done with the project tomorrow, then I may add a auto-rotate feature so the comments show up automatically by themselves. What I need to do by tomorrow is complete the constructor for the about page, then go through and do the media queries for 500, 800, and 1000 pixel widths.